+++ __tests__/ava.js
@@ -1,3 +1,6 @@
+// to avoid ava as deps (since this code is not executed)
+/* eslint-disable import/no-unresolved, import/no-extraneous-dependencies */
+
 import test from "ava"
 
 test("this is a dummy test", (t) => {