+++ __tests__/react-fn.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-unresolved, import/no-extraneous-dependencies */
 import React, { PropTypes } from "react"
 import cx from "classnames"