+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Added: [``eslint-plugin-import``](https://github.com/benmosher/eslint-plugin-import)
+  is now included in the default config.
+  All rules are on except ``no-nodejs-modules`` which depends on the env.
 - Added: [``react/jsx-first-prop-new-line``](https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-first-prop-new-line.md)
   with ``multiline`` value to enforce first prop on a new line if there is
   multiple props.